perf: convert content_offset_seconds from DECIMAL to INTEGER
- Changed database column type from DECIMAL to INTEGER for better storage and indexing
- Updated Prisma schema to use Int instead of Decimal
- Removed unnecessary type conversions (String/parseFloat/Number) throughout codebase
- Updated migrate-streamer script to round values when migrating from legacy DECIMAL
- Improved type safety with TypeScript types matching database reality

+++ prisma/migrations/20260406073608_init/migration.sql
@@ -78,7 +78,7 @@ CREATE TABLE "chat_messages" (
     "id" UUID NOT NULL,
     "vod_id" INTEGER NOT NULL,
     "display_name" TEXT,
-    "content_offset_seconds" DECIMAL NOT NULL,
+    "content_offset_seconds" INTEGER NOT NULL,
     "user_color" TEXT,
     "created_at" TIMESTAMPTZ(6) NOT NULL,
     "message" JSONB,
